requirements.txt
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
streamlit==1.29.0
numpy==1.24.3
pandas==2.1.4
